requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SocketIO==5.3.2
SQLAlchemy==1.4.46
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
python-dotenv==0.21.0
eventlet==0.33.3
requests==2.28.2
gunicorn==20.1.0
email-validator==2.0.0
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.21
stripe==12.0.1
werkzeug==2.3.7
wtforms==3.0.1
